+++ playwright.config.ts
@@ -10,5 +10,5 @@ export default defineConfig({
 	},
 
 	testDir: 'e2e',
-	timeout: 60 * 1000, // 60 seconds
+	timeout: 120 * 1000, // 120 seconds
 })