+++ src/lib/components/Tree/TreeContainer.svelte
@@ -13,10 +13,12 @@
 	import Widgets from './Widgets.svelte'
 	import AddFrames from './AddFrames.svelte'
 	import { useEnvironment } from '$lib/hooks/useEnvironment.svelte'
+	import { usePartID } from '$lib/hooks/usePartID.svelte'
 	const { ...rest } = $props()
 
 	provideTreeExpandedContext()
 
+	const partID = usePartID()
 	const selected = useSelected()
 	const objects = useObjects()
 	const draggable = useDraggable('treeview')
@@ -40,7 +42,7 @@
 </script>
 
 <div
-	class="bg-extralight border-medium absolute top-0 left-0 z-1000 m-2 overflow-y-auto border text-xs"
+	class="bg-extralight border-medium absolute top-0 left-0 z-1000 m-2 w-60 overflow-y-auto border text-xs"
 	style:transform="translate({draggable.current.x}px, {draggable.current.y}px)"
 	{...rest}
 >
@@ -56,9 +58,10 @@
 		/>
 	{/key}
 
-	{#if environment.current.isStandalone}
+	{#if environment.current.isStandalone && partID.current}
 		<AddFrames />
 	{/if}
+
 	<Logs />
 	<Settings />
 	<Widgets />