+++ src/lib/hooks/useFrames.svelte.ts
@@ -6,13 +6,14 @@ import {
 	useResourceNames,
 } from '@viamrobotics/svelte-sdk'
 import { WorldObject, type Geometries } from '$lib/WorldObject.svelte'
-import { observe } from '@threlte/core'
 import { useLogs } from './useLogs.svelte'
 import { resourceColors } from '$lib/color'
-import { usePartConfig, type Frame, type PartConfig } from './usePartConfig.svelte'
+import type { Frame } from '$lib/frame'
+import { usePartConfig, type PartConfig } from './usePartConfig.svelte'
 import { Color } from 'three'
 import { useEnvironment } from './useEnvironment.svelte'
 import { createPoseFromFrame } from '$lib/transform'
+import { observe } from '@threlte/core'
 
 interface FramesContext {
 	current: WorldObject[]
@@ -41,16 +42,13 @@ export const provideFrames = (partID: () => string) => {
 		}
 	)
 
-	observe.pre(
-		() => [partConfig.isDirty],
-		() => {
-			if (partConfig.isDirty) {
-				environment.current.viewerMode = 'edit'
-			} else {
-				environment.current.viewerMode = 'monitor'
-			}
+	$effect.pre(() => {
+		if (partConfig.isDirty) {
+			environment.current.viewerMode = 'edit'
+		} else {
+			environment.current.viewerMode = 'monitor'
 		}
-	)
+	})
 
 	let current = $derived.by(() => {
 		const objects: WorldObject[] = []
@@ -94,15 +92,7 @@ export const provideFrames = (partID: () => string) => {
 			return
 		}
 		worldObject.referenceFrame = component.frame.parent
-		worldObject.localEditedPose = {
-			x: component.frame.translation.x,
-			y: component.frame.translation.y,
-			z: component.frame.translation.z,
-			oX: component.frame.orientation.value.x,
-			oY: component.frame.orientation.value.y,
-			oZ: component.frame.orientation.value.z,
-			theta: component.frame.orientation.value.th,
-		}
+		worldObject.localEditedPose = createPoseFromFrame(component.frame)
 
 		if (component.frame.geometry) {
 			switch (component.frame.geometry.type) {